Apply inheritance to data
